+++ 04.class/command.js
@@ -1,0 +1,21 @@
+const memoController = require("./memoController.js");
+module.exports = Command;
+
+class Command {
+  constructor(option) {
+    this.option = option;
+  }
+
+  runMemo() {
+    if (this.option.l) {
+      return new memoController().index();
+    }
+    if (this.option.r) {
+      return new memoController().reference();
+    }
+    if (this.option.d) {
+      return new memoController().delete();
+    }
+    return new memoController().create();
+  }
+}
